+++ CHANGELOG.md
@@ -9,10 +9,9 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ### Added
 
-- Add `--enable-long-names` feature flag to allow resource names longer than 5 characters. Only for internal testing.
+- Add `--enable-long-names` feature flag to `template cluster/networkpool/nodepool` to allow resource names longer than 5 characters. Only for internal testing.
 
 ## [2.2.0] - 2022-03-04
-
 
 ### Added