+++ CHANGELOG.md
@@ -7,8 +7,11 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
-## [2.2.0] - 2022-03-04
+### Added
 
+- Add `--enable-long-names` feature flag to `template cluster/networkpool/nodepool` to allow resource names longer than 5 characters. Only for internal testing.
+
+## [2.2.0] - 2022-03-04
 
 ### Added